Keep FAQ links on the domain they started on.
Replace `{{ site.url }}` with root-relative paths. This will ensure that
if the user started on graphviz.org, they stay on graphviz.org, rather
than being bounced to graphviz.gitlab.org.

Remove `_pages` from the URL while we're in here - everything in
`/pages/foo` is served out of `/foo`, and the `/` path should be
canonical. `_pages` is an implementation detail of Jekyll that might not
persist with other static site generators, or even future versions of
Jekyll.

Addresses issue #3 for the FAQ page. There are a lot of other pages but
I thought I'd start small.