Remove unused attrs0.md
This just wasn't represented in the output at all.

Grepping for "Testing -" in the output, as is present in the title of
attrs0.md, gave nothing.

And https://graphviz.org/doc/info/attrs0.html 404s.

It had permalink set to attrs.html, but there was already another
attrs.html file clobbering that.

Towards #21