Loop over all gallery pages to generate /Gallery/
Fixes #55

I made the images a bit larger too, now we fit two per column rather
than 4 per column.

I'd have liked each column to be 50% size, but it wrapped for me, so I
hacked it down to 49%, now two columns fit.

None of these style decisions are set in stone, this is just a quick
start to get something running with a loop so it's easier to change in
future.