Add go package dependency example to gallery.
From https://godoc.org/regexp?import-graph

I had to change the code generator in gddo-server to output the raw
graphviz input to get out the raw graphviz:

https://github.com/golang/gddo/blob/721e228c7686d830b5decc691a4dc7e6a6e94888/gddo-server/graph.go#L34

I also changed the URLs to be fully-qualified so the links should work.

gddo-server, which generated this graphviz, is 3-clause-BSD licensed.
But this isn't redistributing source (just debug output) so I think
we're OK to use this as an example.