Add type and defaults to attrs.html headings
The type and defaults is extremely important information for readers.

This saves the reader from jumping back and forth between the table at
the top, and the definition of the attribute.

This was inspired by Pandas Docs:

https://pandas.pydata.org/docs/reference/api/pandas.read_sql_table.html#pandas.read_sql_table

where they have Parameter lists like

"index_col  :  str or list of str, optional, default: None"

With the types and defaults italicised to make clear they are not part
of the main type.