Add graph/cluster/node/edge validity to attrs.html
Saves having to scroll to the table to read where these are appropriate.

I wonder if there's a way to put this nearer the top, perhaps among the 'defaults'/'minimum'? Doesn't seem like there's enough room all the time.

For now, put it at the bottom. We can rethink this later.