Make the huge format image fit the screen width
Removes an HTML validation warning too.

The `%` worked fine before, it's just maybe non-standard HTML?

Anyway, now it fills the same amount of width as the text, which is
better, as we now have a narrow gutter for text in the middle of the
page. 80% before was for a page that was the entire width of the screen,
which isn't quite right.

Long term this should be SVG.