Improve graphs definitions (remove size)
Goal: graphs have to be easy usable with default settings and tools.

Svg generated with "size" definitions has too small viewBox
and the image is located out of the box.
When you open generated svg file with browser the image is out of scope.
Tested with Firefox and Chrome.

Here is diff of svg before and after:

< <svg width="432pt" height="360pt"
<  viewBox="0.00 0.00 432.00 360.11" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
< <g id="graph0" class="graph" transform="scale(3.34 3.34) rotate(0) translate(4 1198)">
---
> <svg width="1442pt" height="1202pt"
>  viewBox="0.00 0.00 1441.96 1202.00" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
> <g id="graph0" class="graph" transform="scale(1 1) rotate(0) translate(4 1198)">