Add separate pages per command-line tool
I'm hoping this can be the beginning of:
1. Having a place to grow more documentation of each tool
2. Being able to link to these tools (not just their manpages)
3. Being able to search these tools and their descriptions using the site search.

A lot of work will still need to be done:
- Converting manpages to HTML
- Adding examples for each tool

But this is a start.

It also helps clear out the top-level "documentation" page and move more into the information hierarchy.

I've put these under "Command Line". Most of these are purely CLI tools, though some have GUIs too. This is a little awkward, but better (I think) than having a section called "tools" -- "tools" is a very generic name.