Add docs and examples for oneblock attribute
I reverse engineered this from the definitions in lib/circogen/circular.c:

```
    if (mapbool(agget(realg, "oneblock")))
  root = createOneBlock(g, &state);
    else
  root = createBlocktree(g, &state);
    circPos(g, root, &state);
```

Seems like this just applies to graphs as it's taking an argument `realg`.

`mapbool` defaults to false.

It's circo-only as it's only found inside lib/circogen.

Figured out the meaning by playing in the playground with an example.

Towards #75