Add docs for undocumented attribute beautify
Docs taken from
https://gitlab.com/graphviz/graphviz/-/commit/ba776ee9271f6af1efec836d4193a621b65b9852
commit message:

> Enable beautify feature in sfdp. This tries to draw leaf nodes
> uniformly on a circle around root node.

Looks like it just applies to graphs as it is `agget(g, "beautify")`.

It's a bool (because of `mapBool` function) which defaults to false.

I've attached examples, but the beautify behaviour seems to have a bug
when viewed in Graphviz web editor: the circle is divided into one too
few sectors, resulting in the first and last node overlapping.

Towards #75