Insruct search engines not to index /_pages/
_pages/ is an accidental URL path that dates back to the Jekyll website
compiler. I have set up redirects from it for backward compatibility in
case anyone was linking to it. It should only be redirects inside
_pages, if there is a “real page” there then something has gone wrong.

Something had indeed gone wrong -- a page was being rendered here.

It seems kind of hard to stop a page being rendered here in the Hugo
API. So we'll do the next best thing: tell Google not to index the page.

This should silence an error we're getting in Google Search Console that
"Duplicate, Google chose different canonical than user".